fix insert return object

+++ EntityRepository.Test/User.cs
@@ -5,7 +5,7 @@ using System.Text;
 namespace EntityRepository.Test {
    public class User2 {
         [PrimaryKey(true)]
-        public short Id { get; set; } //(short, not null)
+        public long Id { get; set; } //(short, not null)
         public byte RoleId { get; set; } //(tinyint, not null)
         public string Username { get; set; } //(varchar(50), not null)
         public string Email { get; set; } //(varchar(100), null)
